refactored exercise types by extracting to own type file
