+++ app/(app)/(tabs)/workouts/[workoutId]/index.tsx
@@ -1,61 +1,15 @@
 import { FontAwesome } from "@expo/vector-icons";
 import { Link, router, useLocalSearchParams } from "expo-router";
 import { useSQLiteContext } from "expo-sqlite/next";
-import {
-  StyleSheet,
-  Text,
-  SafeAreaView,
-  View,
-  Pressable,
-  FlatList,
-} from "react-native";
+import { Text, SafeAreaView, View, Pressable, FlatList } from "react-native";
 import { twColors } from "../../../../../constants/Colors";
-
-// hard coded constants based on the sqlite db table "exercise_type"
-export const exerciseEnums = {
-  RESISTANCE_ENUM: 1,
-  CARDIO_ENUM: 2,
-} as const;
-
-export type ExerciseEnums = typeof exerciseEnums;
-
-export type ExerciseSetParams = {
-  exercise_set_id: number;
-  list_order: number;
-  reps: number;
-  rest_time: number;
-  title: string | null;
-};
-export type ResistanceSetParams = {
-  resistance_set_id: number;
-  total_weight: number;
-};
-export type CardioSetParams = {
-  cardio_set_id: number;
-  target_distance: number | null;
-  target_time: number | null;
-};
-export type UnifiedResistanceSet = ExerciseSetParams & ResistanceSetParams;
-export type UnifiedCardioSet = ExerciseSetParams & CardioSetParams;
-
-// Represent generic exercise with unknown type (i.e unknown if Resistance of Cardio)
-export type ExerciseParams = {
-  exercise_type_id: ExerciseEnums[keyof ExerciseEnums];
-  exercise_id: number;
-  title: string;
-};
-
-type ResistanceSection = {
-  exerciseType: ExerciseEnums["RESISTANCE_ENUM"];
-  exercise: ExerciseParams;
-  data: UnifiedResistanceSet[];
-};
-type CardioSection = {
-  exerciseType: ExerciseEnums["CARDIO_ENUM"];
-  exercise: ExerciseParams;
-  data: UnifiedCardioSet[];
-};
-export type ExerciseSection = ResistanceSection | CardioSection;
+import { ExerciseCard } from "../../../../../components/ExerciseCard";
+import {
+  ExerciseParams,
+  UnifiedCardioSet,
+  UnifiedResistanceSet,
+  exerciseEnums,
+} from "../../../../../utils/exercise-types";
 
 const AddExerciseBtn = ({ workoutId }: { workoutId: string }) => {
   return (
@@ -81,36 +35,6 @@ const AddExerciseBtn = ({ workoutId }: { workoutId: string }) => {
       />
       <Text className="text-2xl text-black dark:text-white">Add Exercise</Text>
     </Pressable>
-  );
-};
-
-const ExerciseCard = ({
-  workoutId,
-  exercise,
-}: {
-  workoutId: number;
-  exercise: {
-    exerciseType: ExerciseEnums[keyof ExerciseEnums];
-    exerciseId: number;
-    sets: UnifiedResistanceSet[] | UnifiedCardioSet[];
-    title: string;
-  };
-}) => {
-  return (
-    <Link
-      asChild
-      style={exerciseStyles.exerciseCard}
-      href={`/(app)/(tabs)/workouts/${workoutId}/${exercise.exerciseId}`}
-    >
-      <Pressable>
-        <Text className="text-3xl font-bold text-black dark:text-white">
-          {exercise.title}
-        </Text>
-        {exercise.exerciseType === exerciseEnums.RESISTANCE_ENUM ?
-          <ResistanceSetList sets={exercise.sets as UnifiedResistanceSet[]} />
-        : <CardioSetList sets={exercise.sets as UnifiedCardioSet[]} />}
-      </Pressable>
-    </Link>
   );
 };
 
@@ -219,44 +143,3 @@ export default function WorkoutDetails() {
     </SafeAreaView>
   );
 }
-
-const CardioSetList = ({ sets }: { sets: UnifiedCardioSet[] }) => {
-  return (
-    <>
-      {sets.map((set) => (
-        <Text className="text-xl dark:text-white">
-          Reps: {set.reps}
-          {"    "}
-          Rest: {set.rest_time}s{"    "}
-          Target Distance:{" "}
-          {set.target_distance ? set.target_distance + "m" : "null"}
-          {"    "}
-          Target Time: {set.target_time ? set.target_time + "s" : "null"}
-        </Text>
-      ))}
-    </>
-  );
-};
-
-const ResistanceSetList = ({ sets }: { sets: UnifiedResistanceSet[] }) => {
-  return (
-    <>
-      {sets.map((set) => (
-        <Text key={set.exercise_set_id} className="text-xl dark:text-white">
-          Reps: {set.reps}
-          {"    "}
-          Rest: {set.rest_time}s{"    "}
-          {set.total_weight}kg
-        </Text>
-      ))}
-    </>
-  );
-};
-
-const exerciseStyles = StyleSheet.create({
-  exerciseCard: {
-    borderBottomWidth: 1,
-    borderBottomColor: twColors.neutral700,
-    padding: 16,
-  },
-});
